Grapples, Ropes & Bridges
Created a rudimentary grappling hook system im considering using for a seperate game mode to not clutter the story gameplay.
Did heaps of experimentation with Verlet intergration and using it to create ropes. Made a rope, sling shot (not collider tho) and  just finished a rope bridge with collision. Still yet to figure out how to make it react to input from collisions but i think I'll get the edge colliders' point and apply force, the rest should be taken care of by the constraints & simulation :smile: